"Fix" unwanted memcpy's
This doesn't fix anything, it might (and it probably does) break a lot
of things since I am completely messing the liveness of data with inside
the air. For now my use case "works" (I didn't really test it that much),
but please DO NOT use that tree, it is utterly broken and not tested.

Signed-off-by: emneo <emneo@kreog.com>